Repair batched-edit losses: shell wiring, codec arm, header, tests
- DhunAppShell: settingsVm creation, drop duplicated call fragment,
  second master + detail call args, TabContent onOpenSettings param
- NavStatePersistence: restore settings decode arm (test needs it)
- LibraryScreen: remove LikedSongsFolderCard splice, apply header entry
- SingleInstanceProtocolTest: drop unresolved token assert
- AndroidEqualizerEngine: force rebind after mid-apply failure
- EqualizerBandMapper: Double.coerceIn needs Double bounds (no mixed
  Float overload); SettingsViewModel doc path fix; ROADMAP refresh

+++ app-android/src/main/kotlin/dev/dhun/android/equalizer/EqualizerBandMapper.kt
@@ -49,9 +49,10 @@ object EqualizerBandMapper {
             val hz = (mhz / 1000.0).coerceAtLeast(1.0)
             val curve = curveGainDb(hz, centers, gains)
             val preamp = state.preampDb.takeUnless { it.isNaN() } ?: 0f
+            // toDouble()s: Double.coerceIn has no mixed Float overload.
             val total = (curve + preamp).coerceIn(
-                EqualizerBands.MIN_GAIN_DB,
-                EqualizerBands.MAX_GAIN_DB,
+                EqualizerBands.MIN_GAIN_DB.toDouble(),
+                EqualizerBands.MAX_GAIN_DB.toDouble(),
             )
             (total * 100).roundToInt().coerceIn(lo, hi).toShort()
         }
